broker/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-blender-broker"
version = "0.1.0"
description = "FastAPI broker for Codex -> Runpod Blender sandbox"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "PyJWT>=2.9.0",
  "requests>=2.32.0",
]

[tool.setuptools]
packages = ["app", "app.routes", "app.services"]
